Create the new controller layer
