+++ config/default.js
@@ -12,6 +12,7 @@ module.exports = {
       '@elifesciences/xpub-meca-export',
       '@elifesciences/xpub-model',
       '@elifesciences/xpub-server',
+      '@elifesciences/xpub-controller',
       '@elifesciences/xpub-client-config',
       '@pubsweet/component-send-email',
     ],
