+++ server/controller/index.js
@@ -1,0 +1,5 @@
+const SupportingFiles = require('./supportingFiles')
+
+module.exports = {
+  SupportingFiles,
+}
