--- server/xpub-server/entities/manuscript/resolvers/test/dummyFile.js
@@ -1,15 +1,0 @@
-const { File } = require('@elifesciences/xpub-model')
-
-class DummyFile extends File {
-  get storage() {
-    return {
-      putContent: () => {},
-      getContent: () => {},
-      deleteObject: () => ({
-        promise: () => Promise.resolve(true),
-      }),
-    }
-  }
-}
-
-module.exports = DummyFile
